+++ service/server.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"context"
 	"net/http"
 	"net/url"
 	"strings"
@@ -13,7 +14,10 @@ import (
 	connectweb "github.com/streamingfast/dgrpc/server/connectrpc"
 	"github.com/streamingfast/dgrpc/server/factory"
 	pbssinternal "github.com/streamingfast/substreams/pb/sf/substreams/intern/v2"
-	ssconnect "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v2/pbsubstreamsrpcconnect"
+	pbsubstreamsrpc "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v2"
+	pbsubstreamsrpcv2connect "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v2/pbsubstreamsrpcv2connect"
+	pbsubstreamsrpcv3 "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v3"
+	pbsubstreamsrpcv3connect "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v3/pbsubstreamsrpcv3connect"
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 	"go.opentelemetry.io/otel"
 	"go.uber.org/zap"
@@ -37,10 +41,16 @@ func GetCommonServerOptions(listenAddr string, logger *zap.Logger, healthcheck d
 	return options
 }
 
+type streamHandlerV3 func(ctx context.Context, req *connect_go.Request[pbsubstreamsrpcv3.Request], stream *connect_go.ServerStream[pbsubstreamsrpc.Response]) error
+
+func (h streamHandlerV3) Blocks(ctx context.Context, req *connect_go.Request[pbsubstreamsrpcv3.Request], stream *connect_go.ServerStream[pbsubstreamsrpc.Response]) error {
+	return h(ctx, req, stream)
+}
+
 func ListenTier1(
 	listenAddr string,
 	svc *Tier1Service,
-	infoService ssconnect.EndpointInfoHandler,
+	infoService pbsubstreamsrpcv2connect.EndpointInfoHandler,
 	auth dauth.Authenticator,
 	logger *zap.Logger,
 	healthcheck dgrpcserver.HealthCheck,
@@ -54,16 +64,23 @@ func ListenTier1(
 
 		options = append(options, dgrpcserver.WithConnectInterceptor(dauthconnect.NewAuthInterceptor(auth, logger)))
 		options = append(options, dgrpcserver.WithConnectStrictContentType(false))
-		options = append(options, dgrpcserver.WithConnectReflection(ssconnect.StreamName))
+		options = append(options, dgrpcserver.WithConnectReflection(pbsubstreamsrpcv2connect.StreamName))
+		options = append(options, dgrpcserver.WithConnectReflection(pbsubstreamsrpcv3connect.StreamName))
 
 		streamHandlerGetter := func(opts ...connect_go.HandlerOption) (string, http.Handler) {
-			return ssconnect.NewStreamHandler(svc, opts...)
+			return pbsubstreamsrpcv2connect.NewStreamHandler(svc, opts...)
 		}
-		handlerGetters := []connectweb.HandlerGetter{streamHandlerGetter}
+
+		streamHandlerGetterV3 := func(opts ...connect_go.HandlerOption) (string, http.Handler) {
+			handler := streamHandlerV3(svc.BlocksV3)
+			return pbsubstreamsrpcv3connect.NewStreamHandler(handler, opts...)
+		}
+
+		handlerGetters := []connectweb.HandlerGetter{streamHandlerGetter, streamHandlerGetterV3}
 
 		if infoService != nil {
 			infoHandlerGetter := func(opts ...connect_go.HandlerOption) (string, http.Handler) {
-				out, outh := ssconnect.NewEndpointInfoHandler(infoService, opts...)
+				out, outh := pbsubstreamsrpcv2connect.NewEndpointInfoHandler(infoService, opts...)
 				return out, outh
 			}
 			handlerGetters = append(handlerGetters, infoHandlerGetter)